Consolidate selector logic into single mpsc module
