+++ src/sync/mod.rs
@@ -7,7 +7,6 @@ pub mod mpsc;
 mod mutex;
 mod once;
 mod rwlock;
-pub mod selector;
 
 pub use barrier::{Barrier, BarrierWaitResult};
 pub use condvar::{Condvar, WaitTimeoutResult};
